requirements.txt
aiofiles==0.7.0
anyio==3.5.0
APScheduler==3.8.1
arrow==1.2.1
asgiref==3.4.1
attrs==21.4.0
binaryornot==0.4.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cookiecutter==1.7.3
fastapi==0.70.1
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.4
httptools==0.2.0
httpx==0.21.3
idna==3.3
importlib-metadata==4.10.0
iniconfig==1.1.1
Jinja2==3.0.3
jinja2-time==0.2.0
loguru==0.5.3
Markdown==3.3.6
MarkupSafe==2.0.1
multidict==5.2.0
nb-cli==0.6.3
nonebot-adapter-cqhttp==2.0.0b1
nonebot-adapter-onebot==2.0.0b1
nonebot-plugin-apscheduler==0.1.2
nonebot-plugin-test==0.3.0
nonebot2==2.0.0b1
packaging==21.3
pluggy==1.0.0
poyo==0.5.0
prompt-toolkit==3.0.24
psutil==5.9.0
py==1.11.0
pydantic==1.9.0
pyee==8.2.2
pyfiglet==0.8.post1
Pygments==2.11.2
pygtrie==2.4.2
pymdown-extensions==9.1
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
python-engineio==3.14.2
python-markdown-math==0.8
python-slugify==5.0.2
python-socketio==4.6.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
text-unidecode==1.3
toml==0.10.2
tomlkit==0.7.2
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.1
yarl==1.7.2
zipp==3.7.0
aiohttp
aiocqhttp
